Fix: `core` method mocks have haste priority
This is due to the way Jest treats mock file imports.
If there are more than one with the same name, they will conflict.

A solution for this is to rename one of the folders (so it won't be
picked out automatically), and manually import the (manual) mock.

More more information see: https://github.com/facebook/jest/issues/2070