Refactor: `SoftwareWallet` now shadows Ethers Wallet
It no longer directly extends the Ethers Wallet Class, but it uses an
instance of it to create the new wallet.

@NOTE We're no longer directly extending the Ethers Wallet Class

This is due to the fact that we need more control over the resulting Class
object (SoftwareWallet in this case).

We're still shadowing the Ethers Wallet, meaning when opening or creating a new
wallet, we will first create a Ethers Wallet instance than pass that along
to the SoftwareWallet constructor.

This way we don't have to deal with non-configurable or non-writable props,
or the providers being baked in.