Refactor: core `userInputValudator` for most use cases
It will now accept only the first argument and validate against that.

Previous, it had expected the whole `arguments` "array" and it would extract the first
argument itself.

This way is much cleaner and it molds itself to most of our use cases where we don't
actually have access to the `arguments` array